Extract `Task` aggregate and repository definitions and move those to `model`.
